#20 created util script and necessary hooks added stub function for category_lookup
